Invert how the kingInDanger method works
Basically, instead of checking all pieces, seeing if they have a king in their path,
we now start with the King and trace paths out to the coords that COULD deliver
check. This results in fewer spaces checked (usually) and also lets us build tables
of the specific coords to check, reducing logic.