Promotions are now emitted as multiple moves ; Also UCI format now includes promotion.
Most other engines treat promotions separately, so this just changes our
own logic to be more in-line. Plus, it simplifies a bunch of logic doing
it this way.

As for the move format change: the optional 5th character for promotion is
valid UCI, I just omitted it accidentally. The optional second param which
is only used sometimes was a sloppy API anyways.